+++ helpers/dailyRays.ts
